#173 Fix: FollowService 구현체/인터페이스 분리 작업
